+++ src/Notices/OnNewChat.cs
@@ -1,9 +1,5 @@
-﻿using System.Collections.Generic;
-using System.Reflection.Emit;
-using HarmonyLib;
+﻿using HarmonyLib;
 using JetBrains.Annotations;
-using Splatform;
-using UnityEngine;
 
 namespace DiscordBot.Notices;
 
@@ -15,14 +11,14 @@ public static class OnNewChat
         [UsedImplicitly]
         private static void Postfix(Talker.Type type, string text)
         {
-            if (!DiscordBotPlugin.ShowChat || !Player.m_localPlayer || type is not Talker.Type.Shout || string.IsNullOrEmpty(text) || text == Localization.instance.Localize("$text_player_arrived")) return;
+            if (!DiscordBotPlugin.ShowChat || type is not Talker.Type.Shout || string.IsNullOrEmpty(text) || text == Localization.instance.Localize("$text_player_arrived")) return;
             switch (DiscordBotPlugin.ChatType)
             {
                 case ChatDisplay.Player:
-                    Discord.instance?.SendMessage(Webhook.Chat, Player.m_localPlayer.GetPlayerName() + $" ({Keys.InGame})", text);
+                    Discord.instance?.SendMessage(Webhook.Chat, (Player.m_localPlayer?.GetPlayerName() ?? ZNet.instance.GetWorldName()) + $" ({Keys.InGame})", text);
                     break;
                 case ChatDisplay.Bot:
-                    Discord.instance?.SendMessage(Webhook.Chat, message: $"{Player.m_localPlayer.GetPlayerName()} {Keys.Shouts} {text.Format(TextFormat.Bold)}");
+                    Discord.instance?.SendMessage(Webhook.Chat, message: $"{Player.m_localPlayer?.GetPlayerName() ?? ZNet.instance.GetWorldName()} {Keys.Shouts} {text.Format(TextFormat.Bold)}");
                     break;
             }
         }