+++ src/Recorder.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
-using System.Threading.Tasks;
 using uGIF;
 using UnityEngine;
 
@@ -15,19 +14,15 @@ public class Recorder : MonoBehaviour
 
     [Header("Discord message")]
     private string playerName = string.Empty;
-    private string message = string.Empty;
+    public string message = string.Empty;
     private string thumbnail = string.Empty;
 
     [Header("GIF Settings")] 
-    private readonly List<Color32[]> recordedFrameData = new();
-
     private readonly List<Image> recordedImages = new();
     
-    // private Texture2D texture = null!;
     private bool isRecording;
     private float recordStartTime;
     private Coroutine? recordingCoroutine;
-    // private RenderTexture? gifTexture;
     private byte[]? gifBytes;
     private static int gifHeight => DiscordBotPlugin.GifResolution.height;
     private static int gifWidth => DiscordBotPlugin.GifResolution.width;
@@ -39,35 +34,11 @@ public class Recorder : MonoBehaviour
     public void Awake()
     {
         instance = this;
-        // gifTexture = new RenderTexture(gifWidth, gifHeight, 24);
-        // gifTexture.Create();
-        // texture = new Texture2D(gifTexture.width, gifTexture.height, TextureFormat.RGB24, false);
     }
     
     public void OnDestroy()
     {
         instance = null;
-    }
-    
-    public void OnGifResolutionChange()
-    {
-        // if (isRecording)
-        // {
-        //     DiscordBotPlugin.LogWarning("Bot is recording GIF, cannot change settings");
-        //     return;
-        // }
-        //
-        // if (gifTexture != null)
-        // {
-        //     gifTexture.Release();
-        //     DestroyImmediate(gifTexture);
-        //     gifTexture = null;
-        // }
-        // gifTexture = new RenderTexture(gifWidth, gifHeight, 24);
-        // gifTexture.Create();
-        //
-        // DestroyImmediate(texture);
-        // texture = new  Texture2D(gifTexture.width, gifTexture.height, TextureFormat.RGB24, false);
     }
     
     public void StartRecording(string player, string quip, string avatar)
@@ -90,7 +61,6 @@ public class Recorder : MonoBehaviour
         while (isRecording && Time.time - recordStartTime < recordDuration)
         {
             yield return new WaitForEndOfFrame();
-            // recordedFrameData.Add(CaptureFrame());
             Image img = new Image(ScreenCapture.CaptureScreenshotAsTexture());
             recordedImages.Add(img);
             yield return new WaitForSeconds(interval);
@@ -112,25 +82,9 @@ public class Recorder : MonoBehaviour
 
     public void Cleanup()
     {
-        // recordedFrameData.Clear();
         recordedImages.Clear();
         gifBytes = null;
     }
-    
-    // private Color32[] CaptureFrame()
-    // {
-    //     RenderTexture previousTarget = camera.targetTexture;
-    //     camera.targetTexture = gifTexture;
-    //     camera.Render();
-    //     RenderTexture.active = gifTexture;
-    //     texture.ReadPixels(new Rect(0, 0, gifWidth, gifHeight), 0, 0, false);
-    //     texture.Apply();
-    //     RenderTexture.active = null;
-    //     camera.targetTexture = previousTarget;
-    //
-    //     texture = ScreenCapture.CaptureScreenshotAsTexture();
-    //     return texture.GetPixels32();
-    // }
     
     private void CreateGif()
     {
@@ -145,15 +99,6 @@ public class Recorder : MonoBehaviour
 
         MemoryStream stream = new MemoryStream();
         encoder.Start(stream);
-
-        // foreach (Color32[]? pixels in recordedFrameData)
-        // {
-        //     Image img = new Image(pixels, size.x, size.y);
-        //     img.ResizeBilinear(gifWidth, gifHeight);
-        //     img.Flip();
-        //     encoder.AddFrame(img);
-        // }
-
         foreach (Image? img in recordedImages)
         {
             img.ResizeBilinear(gifWidth, gifHeight);